squash domaine and sousdomaine entity

+++ src/main/java/fr/uga/miage/m1/repository/DepartementRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartementRepository extends JpaRepository<Departement, Integer> {
+public interface DepartementRepository extends JpaRepository<Departement, Long> {
 }
